requirements.txt
Pillow==11.0.0
scapy==2.6.0
ttkbootstrap==1.10.1
